pyproject.toml
[project]
name = "doc-intelligence-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "mistral>=21.0.0",
    "mistralai>=1.9.11",
    "neo4j>=6.0.2",
    "openai>=2.3.0",
    "pymupdf>=1.26.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "scikit-learn>=1.7.2",
    "tiktoken>=0.12.0",
]
